match sub_8078A58

+++ src/status_actions.c
@@ -34,7 +34,7 @@
 
 extern void sub_807F43C(Entity *, Entity *);
 extern void sub_807DC68(Entity *, Entity *);
-extern void sub_8078A58(Entity *, Entity *, s16, u32);
+extern void sub_8078A58(Entity *, Entity *, s32, u32);
 extern u32 sub_803D73C(u32);
 extern void sub_8045C28(Item *, u8 , u8);
 extern void sub_80464C8(Entity *, Position *, Item *);
